# target/classes/openapi.yaml
openapi: 3.0.0
info:
  description: "Trust FrameWork (TFW) is composed by the parts: MTC, LOA and DataCatalog. My Trusted Companies (MTC) is a module where the entity can configure its rules to establish an advance blacklist of entites which are not considered trusted. Credentials created by them will be refused. Level of Assurance (LoA) manages the level of assurance of the credentials and presentations according to the use case whereas Data Catalog is a corporate system that parametrizes the requested data for the clients. This module is invoked from: JWT Manager and App Integrator Dalion Middleware component."
  title: TRUST FRAMEWORK (TFW) API
  version: 1.0.0
tags:
  - name: demo
    description: Manage the balance operations 
paths:
  /atm/accounts:
    get:
      tags:
        - demo
      summary: Get balance
      operationId: getBalanceAccounts
      parameters:
        - name: id
          in: query
          description: Use case identifier
          required: true
          schema:
            type: integer
            format: int64
        - name: pin
          in: query
          description: pin
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: accounts with balance and dispense limit
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
        "500":
          description: Invalid Order
          content: {}
  /atm/balance:
    get:
      tags:
        - demo
      summary: Get balance
      operationId: getBalance
      parameters:
        - name: id
          in: query
          description: Use case identifier
          required: true
          schema:
            type: integer
            format: int64
        - name: pin
          in: query
          description: pin
          required: true
          schema:
            type: integer
        - name: account
          in: query
          description: pin
          required: true
          schema:
            type: string     
      responses:
        "200":
          description: accounts with balance and dispense limit
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Account'
        "500":
          description: Invalid Order
          content: {} 
  /atm/createAtm:         
    post:
      tags:
        - demo
      summary: Create a group of credentials
      operationId: createAtm
      responses:
        "200":
          description: accounts with balance and dispense limit
        "500":
          description: Invalid Order
          content: {}  
  /atm/withdraw:
    post:
      tags:
        - demo
      summary: Get balance
      operationId: withdrawMoney
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Withdraw'
        description: withdraw money
        required: true     
      responses:
        "200":
          description: accounts with balance and dispense limit
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ammmount'
        "500":
          description: Invalid Order
          content: {}         
       
          
          
components:
  schemas:
    Account:
      type: object
      properties:
        accountId:
          type: string
        balance:
          type: number
          format: float
        maxWithdrawal:
          type: integer
    Withdraw:
      type: object
      properties:
        atmId:
          type: integer
          format: int64
        accountId:
          type: string
        id:
          type: integer
          format: int64  
        pin:
          type: integer
        ammmount:
          type: integer
          multipleOf: 5
            
    Ammmount:
          type: integer
          enum:
            - 5
            - 10
            - 20
            - 50      
                
       